Create 'media' table and update Laravel runtime
Added a new migration file to create a 'media' table in the database. This table will handle media related information like file name, MIME type etc. for the application.

Simultaneously, updated the docker-compose.yml file to use Laravel sail runtime version from 8.1 to 8.2. This is to ensure application's compatibility with the latest version.