"Upgrade Laravel framework and related dependencies
Adjusted Laravel's core framework version from 9.19 to 10.10 in composer.json. This was done to keep the application up to date and take advantage of the latest features and security fixes. Related packages `laravel/sanctum`, `nunomaduro/collision`, `phpunit/phpunit` and `spatie/laravel-ignition` also upgraded.

Meanwhile, the registerPolicies() method was removed in AuthServiceProvider, as it had no content and was unnecessary."