Add BankRelationManager to UserResource
A new BankRelationManager class was added under UserResource's RelationshipManagers to manage the relationship between a user and their bank. The 'getRelations' method was also added to UserResource to include the newly created BankRelationManager. Additionally, 'user_id' is now used to modify the query in 'table' method for user specific data retrieval. These changes aim to streamline the management of user-bank relationships.