Update CSS asset path, sort contracts by date, fix Stripe deposit amount
The path to the CSS in the 'manifest.json' file has been updated after the CSS file was renamed, for maintaining the correct link to the style file. Contracts are now sorted by their 'created_at' date in descending order to make it easier for users to find the most recent ones. Fixed an issue in the StripePaymentSuccessHandlingEvent where the deposit amount was not being divided by 100 resulting in incorrect balance being added to the user's account.