Update contracts, manifest, stripe, & add new resources
In this commit, changes were made primarily to the contract and contract request by adding new conditions to validate payments and contract rules. A change was made in the manifest.json to update the CSS file reference.

Additionally, renamed the StripeRequest to CancelPayoutRequest and included validations for the payout request. A new TransactionResource has been added.

Major changes were done on StripeController by adding multiple methods for handling payout and bank related functionalities. Similarly, new routes in api.php were added to handle bank, payout, and wallet operations in the application.

Also, added new resources for BankDetail, PayoutRequest, and bank related methods on StripeController. Lastly, the CSS file linked in manifest.json was renamed and replaced.

Overall, this commit is aimed to enhance contract handling and payment method validations, and to incorporate bank and payout related functionalities and resources into the app.