Add Stripe payment and transaction confirmation functions
Added two new functions for facilitating Stripe payments and confirming user transactions in the StripeController. A default avatar callback for the user was added in case the fetched avatar is null. Also, flowbite.js was appended to the footer.blade.php for managing UI Flows. Finally, added an artisan command in the web routes file to create a symbolic link for the storage directory. These changes were necessary for improving the transaction operations and the overall user experience in the application.