Refactor contract's status handling logic and add additional data to ContractDetailResource
This commit refactors the status handling in the ContractController. Instead of directly accessing the status properties on a contract, it checks whether a contract status exists first, and if not, it falls back to the corresponding data on the contract. This approach can help to prevent potential null reference exceptions if contract status is not available.

Additionally, seller and buyer proof of contract status have been added to the ContractDetailResource. This change helps to present a more granular view of contract status details to the client.

This change was necessary to improve how the application handles potential null values and it improves status information delivery.