Update condition checks and improve language dropdown
Modified the condition checks in the marketing index blade file for better error handling and updated the language dropdown in the header.blade.php file for a more responsive design. Additionally, added the 'auth:web' middleware to the 'market_details' route in web.php file to ensure user authentication. Added a filter in ProductResource.php for enhanced search functionality.